[REFACTOR-023] feat: extract shared GuidesClient, thin library wrapper, fix GuidesTile hrefs
