+++ app/(dashboard)/components/tiles/GuidesTile.tsx
@@ -35,7 +35,7 @@ export default function GuidesTile({ colSpan = 6, rowSpan, mobileColSpan }: { co
       <div className="flex items-center justify-between mb-4">
         <Eyebrow>{t('home.guides.title')}</Eyebrow>
         <Link
-          href="/guides"
+          href="/library"
           className="text-xs font-semibold tracking-widest uppercase transition-colors hover:opacity-70"
           style={{ color: 'var(--brand-crimson)' }}
         >
@@ -59,7 +59,7 @@ export default function GuidesTile({ colSpan = 6, rowSpan, mobileColSpan }: { co
             return (
               <Link
                 key={g.id}
-                href={`/guides/${g.slug}`}
+                href={`/library/${g.slug}`}
                 className="flex items-center gap-3 px-3 py-2.5 rounded-xl transition-colors hover:bg-black/[0.04] group"
               >
                 <span className="text-base flex-shrink-0 w-6 text-center">
